+++ warp/_src/context.py
@@ -3733,13 +3733,20 @@ class Runtime:
             ]
 
             self.core.wp_bvh_create_host.restype = ctypes.c_uint64
-            self.core.wp_bvh_create_host.argtypes = [ctypes.c_void_p, ctypes.c_void_p, ctypes.c_int, ctypes.c_int]
+            self.core.wp_bvh_create_host.argtypes = [
+                ctypes.c_void_p,
+                ctypes.c_void_p,
+                ctypes.c_int,
+                ctypes.c_int,
+                ctypes.c_int,
+            ]
 
             self.core.wp_bvh_create_device.restype = ctypes.c_uint64
             self.core.wp_bvh_create_device.argtypes = [
                 ctypes.c_void_p,
                 ctypes.c_void_p,
                 ctypes.c_void_p,
+                ctypes.c_int,
                 ctypes.c_int,
                 ctypes.c_int,
             ]
@@ -3761,6 +3768,7 @@ class Runtime:
                 ctypes.c_int,
                 ctypes.c_int,
                 ctypes.c_int,
+                ctypes.c_int,
             ]
 
             self.core.wp_mesh_create_device.restype = ctypes.c_uint64
@@ -3769,6 +3777,7 @@ class Runtime:
                 warp._src.types.array_t,
                 warp._src.types.array_t,
                 warp._src.types.array_t,
+                ctypes.c_int,
                 ctypes.c_int,
                 ctypes.c_int,
                 ctypes.c_int,